+++ stisim/utils.py
@@ -451,6 +451,7 @@ def undefined(arr):
 
 __all__ += ['standardize_data']
 
+
 def standardize_data(data=None, metadata=None):
 
     if isinstance(data, pd.DataFrame):
@@ -476,12 +477,14 @@ def standardize_data(data=None, metadata=None):
             errormsg = 'Could not understand index of data series: should be age (all values less than 120) or year (all values greater than 1900).'
             raise ValueError(errormsg)
 
-        df = pd.concat([df, df])
-        df[metadata.data_cols['sex']] = np.repeat(list(metadata.sex_keys.values()), len(data))
+        if metadata.data_cols.get('sex'):
+            df = pd.concat([df, df])
+            df[metadata.data_cols['sex']] = np.repeat(list(metadata.sex_keys.values()), len(data))
 
     elif isinstance(data, dict):
         if not set(metadata.data_cols.values()).issubset(data.keys()):
             errormsg = 'Please ensure the keys of the data dict match the values in metadata.data_cols.'
+            raise ValueError(errormsg)
         df = pd.DataFrame(data)
 
     elif sc.isnumber(data):