+++ tests/test_demographics.py
@@ -1,5 +1,9 @@
 """
+<<<<<<< HEAD
 Test demographics
+=======
+Test demographic consistency
+>>>>>>> age-distributions
 """
 
 # %% Imports and settings
@@ -7,8 +11,15 @@ import numpy as np
 import stisim as ss
 import pandas as pd
 import matplotlib.pyplot as plt
+<<<<<<< HEAD
+=======
+import numpy as np
+import sciris as sc
+>>>>>>> age-distributions
 
+do_plot = True
 
+<<<<<<< HEAD
 def test_nigeria():
     """ Make a sim with Nigeria demographic data """
 
@@ -34,6 +45,117 @@ def test_nigeria():
             death
         ],
     )
+=======
+
+def test_nigeria(which='births', dt=1, start=1995, n_years=15, plot_init=False, do_plot=True):
+    """
+    Make a Nigeria sim with demographic modules
+    Switch between which='births' or 'pregnancy' to determine which demographic module to use
+    """
+
+    # Make demographic modules
+    demographics = sc.autolist()
+
+    if which == 'births':
+        birth_rates = pd.read_csv(ss.root / 'tests/test_data/nigeria_births.csv')
+        births = ss.births(pars={'birth_rate': birth_rates})
+        demographics += births
+    elif which == 'pregnancy':
+        fertility_rates = pd.read_csv(ss.root / 'tests/test_data/nigeria_asfr.csv')
+        pregnancy = ss.Pregnancy(pars={'fertility_rate': fertility_rates, 'rel_fertility': 1})  # 4/3
+        demographics += pregnancy
+
+    death_rates = pd.read_csv(ss.root / 'tests/test_data/nigeria_deaths.csv')
+    death = ss.background_deaths(pars={'death_rate': death_rates})
+    demographics += death
+
+    # Make people
+    n_agents = 10_000
+    nga_pop_1995 = 106819805
+    age_data = pd.read_csv(ss.root / 'tests/test_data/nigeria_age.csv')
+    ppl = ss.People(n_agents, age_data=age_data)
+
+    sim = ss.Sim(
+        dt=dt,
+        total_pop=nga_pop_1995,
+        start=start,
+        n_years=n_years,
+        people=ppl,
+        demographics=demographics,
+    )
+
+    if plot_init:
+        sim.initialize()
+        # Plot histograms of the age distributions - simulated vs data
+        bins = np.arange(0, 101, 1)
+        init_scale = nga_pop_1995 / n_agents
+        counts, bins = np.histogram(sim.people.age, bins)
+        plt.bar(bins[:-1], counts * init_scale, alpha=0.5, label='Simulated')
+        plt.bar(bins, age_data.value.values * 1000, alpha=0.5, color='r', label='Data')
+        plt.legend(loc='upper right')
+        plt.show()
+
+    sim.run()
+
+    end = start + n_years
+    nigeria_popsize = pd.read_csv(ss.root / 'tests/test_data/nigeria_popsize.csv')
+    data = nigeria_popsize[(nigeria_popsize.year >= start) & (nigeria_popsize.year <= end)]
+
+    nigeria_cbr = pd.read_csv(ss.root / 'tests/test_data/nigeria_births.csv')
+    cbr_data = nigeria_cbr[(nigeria_cbr.Year >= start) & (nigeria_cbr.Year <= end)]
+
+    nigeria_cmr = pd.read_csv(ss.root / 'tests/test_data/nigeria_cmr.csv')
+    cmr_data = nigeria_cmr[(nigeria_cmr.Year >= start) & (nigeria_cmr.Year <= end)]
+
+    # Tests
+    if which == 'pregnancy':
+
+        print("Check we don't have more births than pregnancies")
+        assert sum(sim.results.pregnancy.births) <= sum(sim.results.pregnancy.pregnancies)
+        print(f'✓ (births <= pregnancies)')
+
+        if dt == 1:
+            print("Checking that births equal pregnancies with dt=1")
+            assert np.array_equal(sim.results.pregnancy.pregnancies, sim.results.pregnancy.births)
+            print(f'✓ (births == pregnancies)')
+
+    print("Check final pop size within 5% of data")
+    assert np.isclose(data.n_alive.values[-1], sim.results.n_alive[-1], rtol=0.05)
+    print(f'✓ (simulated/data={sim.results.n_alive[-1] / data.n_alive.values[-1]:.2f})')
+
+    # Plots
+    if do_plot:
+        fig, ax = plt.subplots(2, 2)
+        ax = ax.ravel()
+        ax[0].scatter(data.year, data.n_alive, alpha=0.5)
+        ax[0].plot(sim.yearvec, sim.results.n_alive, color='k')
+        ax[0].set_title('Population')
+
+        ax[1].plot(sim.yearvec, 1000 * sim.results.background_deaths.cmr / dt, label='Simulated CMR')
+        ax[1].scatter(cmr_data.Year, cmr_data.CMR, label='Data CMR')
+        ax[1].set_title('CMR')
+        ax[1].legend()
+
+        if which == 'births':
+            ax[2].plot(sim.yearvec, sim.results.births.cbr / dt, label='Simulated CBR')
+        elif which == 'pregnancy':
+            ax[2].plot(sim.yearvec, sim.results.pregnancy.cbr / dt, label='Simulated CBR')
+        ax[2].scatter(cbr_data.Year, cbr_data.CBR, label='Data CBR')
+        ax[2].set_title('CBR')
+        ax[2].legend()
+
+        if which == 'pregnancy':
+            ax[3].plot(sim.yearvec, sim.results.pregnancy.pregnancies / dt, label='Pregnancies')
+            ax[3].plot(sim.yearvec, sim.results.pregnancy.births / dt, label='Births')
+            ax[3].set_title('Pregnancies and births')
+            ax[3].legend()
+
+        fig.tight_layout
+
+        plt.show()
+
+    return sim
+>>>>>>> age-distributions
 
     sim.run()
 
@@ -57,6 +179,7 @@ def test_nigeria():
 
     fig.tight_layout
 
+<<<<<<< HEAD
     plt.show()
 
     return sim
@@ -66,3 +189,8 @@ if __name__ == '__main__':
 
     sim = test_nigeria()
 
+=======
+if __name__ == '__main__':
+    # Test Nigeria demographic consistency
+    sim = test_nigeria(dt=1, which='pregnancy', n_years=15, plot_init=False, do_plot=do_plot)
+>>>>>>> age-distributions